split trends types into their own file
